+++ include/nstd/gl/texture/texture.h
@@ -6,28 +6,8 @@
 #include "../surface/surface.h"
 NSTDCPPSTART
 
-/// A surface texture.
-typedef NSTDAny NSTDGLSurfaceTexture;
-
 /// A texture format.
 typedef NSTDAny NSTDGLTextureFormat;
-
-/// Gets a surface's current texture.
-/// Parameters:
-///     `const NSTDGLSurface surface` - The surface.
-/// Returns: `NSTDGLSurfaceTexture surface_texture` - The surface's texture.
-NSTDAPI NSTDGLSurfaceTexture nstd_gl_surface_texture_current(const NSTDGLSurface surface);
-
-/// Presents a surface texture to the surface.
-/// NOTE: This function will free `surface_texture`.
-/// Parameters:
-///     `NSTDGLSurfaceTexture *const surface_texture` - The surface texture to present.
-NSTDAPI void nstd_gl_surface_texture_present(NSTDGLSurfaceTexture *const surface_texture);
-
-/// Frees a surface texture.
-/// Parameters:
-///     `NSTDGLSurfaceTexture *const surface_texture` - The surface texture to free.
-NSTDAPI void nstd_gl_surface_texture_free(NSTDGLSurfaceTexture *const surface_texture);
 
 /// Gets the surface's preferred format.
 /// Parameters: